// frontend/certifications-ui/src/app/core/api/generated/certificationsApiV1.schemas.ts
export interface CertificationDto {
  certificationId: number;
  contractId: number;
  assessor: string;
  certificationDate: string;
  /** @nullable */
  protocolDate?: string | null;
  /** @nullable */
  prolongationSend?: string | null;
  /** @nullable */
  prolongationReturned?: string | null;
  isCompleted: boolean;
}
// ...
export type ContractDtoStatus = (typeof ContractDtoStatus)[keyof typeof ContractDtoStatus];

export const ContractDtoStatus = {
  NotApplicable: 'NotApplicable',
  ContractValid: 'ContractValid',
  CertificationPending: 'CertificationPending',
  CertificationInProgress: 'CertificationInProgress',
  CertificationMissing: 'CertificationMissing',
} as const;

export interface ContractDto {
  contractId: number;
  employeeId: string;
  position: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  division?: string | null;
  contractDate: string;
  /** @nullable */
  validTo?: string | null;
  effectiveValidTo: string;
  active: boolean;
  prolongationWarningMonths: number;
  prolongationAlertMonths: number;
  prolongationForYears: number;
  rowVersion: number;
  status: ContractDtoStatus;
}
// ...
export interface UpdateCertificationRequest {
  assessor: string;
  certificationDate: string;
  /** @nullable */
  protocolDate?: string | null;
  /** @nullable */
  prolongationSend?: string | null;
}